# perf: replace deepmerge npm-package with a more performant implementation