@mcollina 2022-06-26
I think we could use a faster version of that modules in a couple more places.

Maybe we should put it in a module? cc @fastify/core