@dalisoft 2022-06-26
@mcollina I am vote for make modules within @fastify organization to separate module and keep this module readable