@Uzlopak 2022-06-27
If you want to have it in a separate module, then I would recommend to create a new project in the org, like we did with the accept-negotatior ;).